+++ bliss/controllers/mca/handel/interface.py
@@ -8,8 +8,12 @@ import logging
 from functools import reduce
 
 import numpy
+import time
 import gevent
 from bliss.common import event
+from bliss.config import settings
+from bliss.config.static import get_config as get_beacon_config
+from bliss.config.beacon_object import BeaconObject
 
 from .error import check_error, HandelError
 from ._cffi import handel, ffi
@@ -20,7 +24,7 @@ from .mapping import parse_mapping_buffer
 __all__ = [
     "init",
     "init_handel",
-    "exit",
+    "exit_handel",
     "get_num_detectors",
     "get_detectors",
     "get_detector_from_channel",
@@ -122,23 +126,82 @@ def init(*path):
     """ Called at BLISS object init.
     """
     filename = to_bytes(os.path.join(*path))
-    LOGGER.debug("init (config={%s})", filename)
+    LOGGER.info("init (config={%s})", filename)
 
     code = handel.xiaInit(filename)
     check_error(code)
 
 
-def init_handel():
+def init_handel(mca_name):
     """ Called at server startup.
+    Do:
+    - xiaInit()
+    - xiaLoadSystem()
     """
-    LOGGER.debug("init_handel --")
-
     code = handel.xiaInitHandel()
     check_error(code)
 
+    # vmaj, vmin, vrel = get_handel_version()  # exit prog ???
+    # vmaj, vmin, vrel = 0, 0, 0
+    # LOGGER.info("init_handel -- %s -- version=%s.%s.%s", mca_name, vmaj, vmin, vrel)
 
-def exit():
-    LOGGER.debug("exit()")
+    # Get info from Beacon configuration and settings.
+    print(f"get config for {mca_name}")
+    cfg = get_beacon_config()
+    mca_config = cfg.get_config(mca_name)
+    if mca_config is None:
+        raise ValueError(f"Cannot find config for {mca_name}")
+    mca_beacon_obj = BeaconObject(mca_config)
+
+    # Get Beacon setting parameters.
+    current_configuration = mca_beacon_obj.settings.get("current_configuration")
+    LOGGER.debug("current_configuration = %s", current_configuration)
+
+    # Get Beacon configuration parameters.
+    default_configuration = mca_config.get("default_configuration")
+    LOGGER.debug("default_configuration = %s", default_configuration)
+    config_dir = mca_config.get("configuration_directory")
+    LOGGER.debug("config_dir = %s", config_dir)
+
+    # Use default_configuration if current_configuration is not found.
+    if current_configuration is None:
+        print("No 'current_configuration' found in settings")
+        try:
+            init(config_dir, default_configuration)
+            print(f"Default configuration {default_configuration} loaded.")
+            current_configuration = default_configuration
+        except Exception:
+            raise RuntimeError(
+                f"Error loading default configuration {default_configuration}"
+            )
+    else:
+        try:
+            # Load current_configuration.
+            init(config_dir, current_configuration)
+            print(f"Current configuration {current_configuration} loaded.")
+        except Exception:
+            # Load 'default_configuration' in case of failure of
+            # current_configuration loading (incorrect config file for example).
+            print(f"Loading current configuration {current_configuration} failed.")
+            try:
+                init(config_dir, default_configuration)
+                print(f"Default configuration {default_configuration} loaded.")
+                current_configuration = default_configuration
+            except Exception:
+                raise RuntimeError(
+                    f"Error loading default configuration {default_configuration}"
+                )
+
+    # In case of success, update 'current_configuration' Beacon setting.
+    mca_beacon_obj.settings.set({"current_configuration": current_configuration})
+    print(f"current config is now: {current_configuration}")
+
+    start_system()
+    LOGGER.debug("end of init_handel()")
+
+
+def exit_handel():
+    LOGGER.debug("exit_handel()")
     code = handel.xiaExit()
     check_error(code)
 
@@ -194,19 +257,18 @@ def start_run(channel=None, resume=False):
 def stop_run(channel=None):
     """
     """
-    LOGGER.debug(f"stop_run(channel={channel})")
-
     if channel is None:
         channel = -1  # All channels
+    LOGGER.debug(f"stop_run(channel={channel})")
     code = handel.xiaStopRun(channel)
     check_error(code)
 
 
 def get_channel_realtime(channel):
-    time = ffi.new("double *")
-    code = handel.xiaGetRunData(channel, b"realtime", time)
+    timing = ffi.new("double *")
+    code = handel.xiaGetRunData(channel, b"realtime", timing)
     check_error(code)
-    return time[0]
+    return timing[0]
 
 
 def get_spectrum_length(channel):
@@ -217,7 +279,6 @@ def get_spectrum_length(channel):
 
 
 def get_spectrum(channel):
-    LOGGER.debug(f"get_spectrum(channel={channel})")
     length = get_spectrum_length(channel)
     array = numpy.zeros(length, dtype="uint32")
     data = ffi.cast("uint32_t *", array.ctypes.data)
@@ -228,6 +289,7 @@ def get_spectrum(channel):
 
 def get_spectrums():
     """Return the spectrums for all enabled channels as a dictionary."""
+    LOGGER.debug("get_spectrums()")
     return {channel: get_spectrum(channel) for channel in get_channels()}
 
 
@@ -251,9 +313,11 @@ def is_running():
 
     if LOGGER.level == logging.DEBUG:
         if running:
-            print("R ", end="")
+            pass
+            # print("\rRunning ", end="")
         else:
-            print("Not Running")
+            pass
+            # print("Not Running")
 
     return running
 
@@ -297,7 +361,7 @@ def get_module_statistics(module):
     # Run handel call
     code = handel.xiaGetRunData(master, b"module_statistics_2", data)
     check_error(code)
-    # Parse raw data
+    LOGGER.info("Parse raw data")
     return {
         channel: stats_from_normal_mode(array[index * 9 :])
         for index, channel in enumerate(channels)
@@ -307,16 +371,17 @@ def get_module_statistics(module):
 
 def get_statistics():
     """Return the statistics for all enabled channels as a dictionary."""
+    LOGGER.debug("get_statistics()")
     result = {}
     # We're not using get_master_channels here.
     # That's because each FalconX channels is its own master, even though
     # the statistics can be accessed with a single call per module.
     for module in get_modules():
-        result.update(get_module_statistics(module))
+        stats = get_module_statistics(module)
+        LOGGER.debug("stats=%r", stats)
+        result.update(stats)
 
-    LOGGER.debug("get_statistics()")
-    LOGGER.debug("result=", result)
-
+    # print(result)
     return result
 
 
@@ -328,7 +393,7 @@ def get_buffer_length(master):
     code = handel.xiaGetRunData(master, b"buffer_len", length)
     check_error(code)
     buf_len = length[0]
-    LOGGER.debug(f"get_buffer_length(master={master}) --> {buf_len}")
+    # LOGGER.debug(f"get_buffer_length(master={master}) --> {buf_len}")
 
     return buf_len
 
@@ -341,7 +406,7 @@ def is_buffer_full(master, buffer_id):
     check_error(code)
     ibf = bool(result[0])
 
-    LOGGER.debug(f"is_buffer_full(master={master}, buffer_id={buffer_id}) --> {ibf}")
+    # LOGGER.debug(f"is_buffer_full(master={master}, buffer_id={buffer_id}) --> {ibf}")
 
     return ibf
 
@@ -351,17 +416,22 @@ def is_buffer_overrun(master):
     code = handel.xiaGetRunData(master, b"buffer_overrun", result)
     check_error(code)
     ibo = bool(result[0])
-    LOGGER.debug(f"is_buffer_overrun(master={master}) --> {ibo}")
+    if ibo:
+        LOGGER.debug(f"is_buffer_overrun(master={master}) --> {ibo}")
     return ibo
 
 
 def get_raw_buffer(master, buffer_id):
-    LOGGER = logging.getLogger("HANDEL_rpc")
-    LOGGER.debug(f"get_raw_buffer(master={master} buff id={buffer_id})")
-
+    """
+    Read currently acquiered data from handel.
+    Return raw data.
+    Used in mapping mode.
+    """
     bid = to_buffer_id(buffer_id)
     command = b"buffer_%c" % bid
     length = get_buffer_length(master)
+    LOGGER.debug(f"get_raw_buffer(master={master} buff id={buffer_id}) (len={length})")
+
     array = numpy.zeros(length * 2, dtype="uint16")
     data = ffi.cast("uint32_t *", array.ctypes.data)
     code = handel.xiaGetRunData(master, command, data)
@@ -376,7 +446,7 @@ def get_raw_buffer(master, buffer_id):
 
 
 def get_buffer_data(master, buffer_id):
-    LOGGER.debug(f"get_buffer_date(master={master} buff id={buffer_id})")
+    LOGGER.debug(f"get_buffer_data(master={master} buff id={buffer_id})")
     raw = get_raw_buffer(master, buffer_id)
     return parse_mapping_buffer(raw)
 
@@ -404,7 +474,7 @@ def set_buffer_done(master, buffer_id):
     other = b"b" if bid == b"a" else b"a"
     overrun = is_buffer_full(master, other) and is_channel_running(master)
     LOGGER.debug(
-        f"get_raw_buffer(master={master} buff id={buffer_id} )   overrun={overrun}"
+        f"set_buffer_done(master={master} buff id={buffer_id} )   overrun={overrun}"
     )
     return overrun
 
@@ -416,7 +486,7 @@ def set_maximum_pixels_per_buffer():
     """Set the maximum number of pixels per buffer.
 
     It makes sure all the modules are configured with the same value,
-    in order to be able to perform synchronized run.
+    to be able to perform synchronized run.
     """
     set_acquisition_value("num_map_pixels_per_buffer", -1)
     value = min(
@@ -452,11 +522,17 @@ def set_all_buffer_done(buffer_id):
 
 
 def get_current_pixel():
-    """Get the current pixel reported by the hardware."""
+    """
+    Return the max over all current pixel's channel.
+    """
 
-    current_pixel = max(
+    # why MAX ???
+
+    all_current_pixels = (
         get_buffer_current_pixel(master) for master in get_master_channels()
     )
+    current_pixel = max(all_current_pixels)
+    # current_pixel = min(all_current_pixels)
 
     LOGGER.debug(f"get_current_pixel() ->  {current_pixel}")
 
@@ -475,7 +551,9 @@ def get_all_buffer_data(buffer_id):
     return merge_buffer_data(*data)
 
 
-def synchronized_poll_data(done=set()):
+def synchronized_poll_data(
+    acquisition_number, done=set(), pixel_seen_cache={"pixel": 0, "times": 0}
+):
     """Convenient helper for buffer management in mapping mode.
 
     It assumes that all the modules are configured with the same number
@@ -495,32 +573,59 @@ def synchronized_poll_data(done=set()):
     first indexed by pixel and then by channel. If there is no data to
     report, those values are empty dicts.
     """
+    pixel_seen = pixel_seen_cache
     data = {"a": None, "b": None}
-    overrun_error = RuntimeError("Buffer overrun!")
+
+    overrun_error_hwd = RuntimeError("Buffer overrun (hwd)!")
+    overrun_error_soft = RuntimeError("Buffer overrun (soft)!")
     # Get info from hardware
     current_pixel = get_current_pixel()
+
+    # put "a" or "b" in "full" if buffer a or buffer b is full.
     full = {x for x in data if all_buffer_full(x)}  # <- full is a set, not a dict...
-    # Overrun from hardware
+    # Check overrun detected by hardware.
     if any_buffer_overrun():
-        raise overrun_error
+        raise overrun_error_hwd
+
     # FalconX hack
-    # The buffer_done command does not reset the full flag.
+    # The 'buffer_done' command does not reset the full flag.
     # It's only reset when the buffer starts being filled up again.
     # For this reason, we need to remember full flags from the previous call.
-    # This is exactly what the done set does.
+    # This is exactly what the 'done' set does.
     done &= full  # Reset done flags
     full -= done  # Don't read twice
     done |= full  # Set done flags
+
     # Read data from buffers
     for x in full:
         data[x] = get_all_buffer_data(x)
-        # Overrun from set_buffer_done
+        # mark buffers as read. get overrun from set_buffer_done.
         if set_all_buffer_done(x):
-            raise overrun_error
+            raise overrun_error_soft
+
     # Extract data
     args = filter(None, data.values())
     spectrums, stats = merge_buffer_data(*args)
-    # Return
+
+    nb_spectrums = len(spectrums)
+
+    # Count number of times a specific pixel is seen.
+    if pixel_seen["pixel"] == current_pixel:
+        pixel_seen["times"] += 1
+    else:
+        pixel_seen["times"] = 0
+        pixel_seen["pixel"] = current_pixel
+
+    pxs = pixel_seen["pixel"]
+    pxtimes = pixel_seen["times"]
+    missing = acquisition_number - current_pixel
+    if pixel_seen["times"] > 50:
+        if pixel_seen["times"] % 10 == 0:
+            print(
+                f"\rpixel {pxs} seen {pxtimes:3d} times (nb_spectrums={nb_spectrums}) (missing={missing})",
+                end="",
+            )
+
     return current_pixel, spectrums, stats
 
 
@@ -572,6 +677,8 @@ def save_system(*path):
 
 def start_system():
     """
+    Prepare hardware specified in .ini file.
+    To be called once after loading of a .ini file.
     """
     LOGGER.debug("start_system()")
 
@@ -630,6 +737,9 @@ def get_num_modules():
 
 
 def get_modules():
+    """ Return alias(es) of modules.
+    Ex: ['module1']
+    """
     n = get_num_modules()
     arg = [ffi.new("char []", MAX_STRING_LENGTH) for _ in range(n)]
     code = handel.xiaGetModules(arg)
@@ -691,7 +801,9 @@ def get_module_channel_at(alias, index):
 
 
 def get_module_channels(alias):
-    """Return the module channels properly indexed."""
+    """Return the module channels properly indexed.
+    Ex: fx2._proxy.get_module_channels(b"module1")  -> [0, 1]
+    """
     number_of_channels = get_module_number_of_channels(alias)
     return tuple(
         get_module_channel_at(alias, index) for index in range(number_of_channels)
@@ -699,8 +811,12 @@ def get_module_channels(alias):
 
 
 def get_grouped_channels():
-    """Return the indexed channels grouped by modules."""
-    return tuple(get_module_channels(alias) for alias in get_modules())
+    """Return the indexed channels grouped by modules.
+    Ex: [[0, 1]]  (list in BLISS after rpc transport.)
+    """
+    grp_chans = tuple(get_module_channels(alias) for alias in get_modules())
+    # print("grp_chans= ", grp_chans)
+    return grp_chans
 
 
 def get_channels():
@@ -741,7 +857,7 @@ def get_trigger_channels():
 # int xiaGetModules_VB(unsigned int index, char *alias);
 
 
-# Parameters
+# Parameters (acquisition values)
 
 
 def get_acquisition_value(name, channel=None):
@@ -767,6 +883,7 @@ def get_acquisition_value(name, channel=None):
 
 
 def set_acquisition_value(name, value, channel=None):
+    # LOGGER.info(f"set_acquisition_value({name}, \t{value}, ch={channel})")
     if channel is None:
         channel = -1  # All channels
     name = to_bytes(name)
@@ -796,7 +913,60 @@ def apply_acquisition_values(channel=None):
     check_error(code)
 
 
+def print_acq_vals():
+    av_list = [
+        "analog_gain",
+        "analog_offset",
+        "detector_polarity",
+        "detector_polarity",
+        "termination",
+        "attenuation",
+        "coupling",
+        "decay_time",
+        #    "dc_offset",
+        "reset_blanking_enable",
+        #    "reset_blanking_threshold",
+        "reset_blanking_presamples",
+        "reset_blanking_postsamples",
+        "clock_speed",
+        "adc_trace_decimation",
+        #    "detection_threshold",
+        "min_pulse_pair_separation",
+        #    "risetime_optimization",
+        "detection_filter",
+        #    "scale_factor",
+        "number_mca_channels",
+        "mca_spectrum_accepted",
+        "mca_spectrum_rejected",
+        "mca_start_channel",
+        "mca_refresh",
+        "mca_bin_width",
+        "preset_type",
+        "preset_value",
+        "mapping_mode",
+        "sca_trigger_mode",
+        "sca_pulse_duration",
+        "number_of_scas",
+        #    "sca",
+        "num_map_pixels",
+        "num_map_pixels_per_buffer",
+        "pixel_advance_mode",
+        "input_logic_polarity",
+        "gate_ignore",
+        "sync_count",
+        "auto_dc_offset",
+    ]
+
+    for av in av_list:
+        value = get_acquisition_value(av)
+        print(f"{av} = {value}")
+
+
 def trigger():
+    """
+    Called only in MCA mode (step by step scan).
+    """
+    LOGGER.debug("trigger()")
     module = inspect.getmodule(trigger)
     try:
         stop_run()
@@ -829,9 +999,14 @@ def wait_hardware_reading():
             return HARDWARE_READING_TASK.get()
         finally:
             HARDWARE_READING_TASK = None
+    print("\nEnd of HARDWARE_READING_TASK\n")
 
 
 def _hardware_poll_points(npoints):
+    """
+    Read data from handel lib and put them in a queue.
+    * spawned from _hardware_poll_points()
+    """
     module = inspect.getmodule(_hardware_poll_points)
     queue = gevent.queue.Queue()
     try:
@@ -850,6 +1025,7 @@ def _hardware_poll_points(npoints):
 def _raw_read(acquisition_number, queue):
     module = inspect.getmodule(_raw_read)
     send_pixel_task = None
+    LOGGER.debug(f"_raw_read(acquisition_number={acquisition_number})")
     try:
         current_pixel_event = gevent.event.Event()
         current_pixel_dict = {"current": -1}
@@ -861,31 +1037,52 @@ def _raw_read(acquisition_number, queue):
                 event.send(module, "current_pixel", current_pixel_dict["current"])
 
         def poll_data(sent):
-            current, data, statistics = synchronized_poll_data()
+            current, data, statistics = synchronized_poll_data(acquisition_number)
             points = list(range(sent, sent + len(data)))
+
+            # To inform BLISS about progression.
             current_pixel_dict["current"] = current
             current_pixel_event.set()
+
             # Check data integrity
             if sorted(data) != sorted(statistics) != points:
                 raise RuntimeError("The polled data overlapped during the acquisition")
             sent += len(data)
+
             # Send the data
             for n in points:
                 queue.put((data[n], statistics[n]))
+            print(
+                f"\r in poll_data current_pixel={current} sent={sent}/{acquisition_number}",
+                end="",
+            )
+
             # Finished
             # we should go in this test to send the end of the acquisition
             if sent == current == acquisition_number:
+                print("")
                 raise StopIteration
-            # Sleep
+
             gevent.sleep(0)
             return sent
 
         send_pixel_task = gevent.spawn(send_current_pixel)
         sent = 0
         while is_running():
+            _t0 = time.time()
             sent = poll_data(sent)
+            _duration = time.time() - _t0
+            LOGGER.debug(
+                f"poll_data() after poll_data ({_duration}) sent={sent} acq_nb={acquisition_number}"
+            )
+
         # get last points
+        _t0 = time.time()
+        _duration = time.time() - _t0
         poll_data(sent)
+        LOGGER.debug(
+            f"poll_data() after poll_data ({_duration}) sent={sent} acq_nb={acquisition_number}"
+        )
     except StopIteration:
         pass
     except Exception as e:
@@ -917,17 +1114,16 @@ def _raw_read(acquisition_number, queue):
 # int xiaCommandOperation(int detChan, byte_t cmd, unsigned int lenS,
 #                         byte_t *send, unsigned int lenR, byte_t *recv);
 
-
 # Debugging
 
 
 def get_handel_version():
-    rel = ffi.new("int *")
-    min = ffi.new("int *")
-    maj = ffi.new("int *")
-    pretty = ffi.new("char *")
-    handel.xiaGetVersionInfo(rel, min, maj, pretty)
-    return maj[0], min[0], rel[0]
+    _rel = ffi.new("int *")
+    _min = ffi.new("int *")
+    _maj = ffi.new("int *")
+    _pretty = ffi.new("char *")
+    handel.xiaGetVersionInfo(_rel, _min, _maj, _pretty)
+    return _maj[0], _min[0], _rel[0]
 
 
 # Not exposed