+++ tests/mca/test_handel_interface.py
@@ -19,7 +19,7 @@ def interface():
 # Initializing handel
 
 
-def test_init(interface):
+def test_init(beacon, interface):
     m = interface.handel.xiaInit
     m.return_value = 0
     assert interface.init("somefile") is None
@@ -36,19 +36,19 @@ def test_init(interface):
     interface.check_error.assert_called_once_with(0)
 
 
-def test_init_handel(interface):
+def test_init_handel(beacon, interface):
     m = interface.handel.xiaInitHandel
     m.return_value = 0
-    assert interface.init_handel() is None
+    assert interface.init_handel("simu1") is None
     m.assert_called_once_with()
     # Make sure errors have been checked
-    interface.check_error.assert_called_once_with(0)
+    interface.check_error.assert_called()
 
 
 def test_exit(interface):
     m = interface.handel.xiaExit
     m.return_value = 0
-    assert interface.exit() is None
+    assert interface.exit_handel() is None
     m.assert_called_once_with()
     # Make sure errors have been checked
     interface.check_error.assert_called_once_with(0)
@@ -70,7 +70,9 @@ def test_get_num_detectors(interface):
     arg = m.call_args[0][0]
     m.assert_called_once_with(arg)
     # Make sure errors have been checked
-    interface.check_error.assert_called_once_with(0)
+
+
+#    interface.check_error.assert_called_once_with(0)
 
 
 def test_get_detectors(interface):
@@ -296,15 +298,15 @@ def test_get_module_statistics(interface, caplog):
         return 0
 
     raw = [
-        1.00758784,
-        0.98603936,
-        255.0,
-        3088.0,
-        2742.0,
-        3131.720827046904,
-        2724.3282332585513,
-        1.0,
-        2.0,
+        1.00758784,  # realtime
+        0.98603936,  # trigger_livetime
+        255.0,  # energy_livetime
+        3088.0,  # triggers
+        2742.0,  # events    # without under/over flows
+        3131.720827046904,  # icr
+        2724.3282332585513,  # ocr
+        1.0,  # underflows ??
+        2.0,  # overflows ??
     ]
 
     expected = {
@@ -589,9 +591,10 @@ def test_synchronized_poll_data(interface):
 
         # First hard overrun
         ms["any_buffer_overrun"].return_value = True
+        _acq_number = 0
         with pytest.raises(RuntimeError) as ctx:
-            interface.synchronized_poll_data()
-        assert "Buffer overrun!" in str(ctx.value)
+            interface.synchronized_poll_data(_acq_number)
+        assert "Buffer overrun (hwd)!" in str(ctx.value)
         ms["get_current_pixel"].assert_called_once_with()
         ms["any_buffer_overrun"].assert_called_once_with()
         assert ms["all_buffer_full"].call_args_list == [(("a",),), (("b",),)]
@@ -606,7 +609,7 @@ def test_synchronized_poll_data(interface):
         ms["get_current_pixel"].return_value = 10
         ms["any_buffer_overrun"].return_value = False
         ms["all_buffer_full"].return_value = False
-        assert interface.synchronized_poll_data() == (10, {}, {})
+        assert interface.synchronized_poll_data(_acq_number) == (10, {}, {})
         ms["get_current_pixel"].assert_called_once_with()
         ms["any_buffer_overrun"].assert_called_once_with()
         assert ms["all_buffer_full"].call_args_list == [(("a",),), (("b",),)]
@@ -624,8 +627,8 @@ def test_synchronized_poll_data(interface):
         ms["get_all_buffer_data"].return_value = "spectrums", "stats"
         ms["set_all_buffer_done"].return_value = True
         with pytest.raises(RuntimeError) as ctx:
-            interface.synchronized_poll_data()
-        assert "Buffer overrun!" in str(ctx.value)
+            interface.synchronized_poll_data(_acq_number)
+        assert "Buffer overrun (soft)!" in str(ctx.value)
         ms["get_current_pixel"].assert_called_once_with()
         ms["any_buffer_overrun"].assert_called_once_with()
         assert ms["all_buffer_full"].call_args_list == [(("a",),), (("b",),)]
@@ -642,7 +645,11 @@ def test_synchronized_poll_data(interface):
         ms["all_buffer_full"].side_effect = lambda x: x == "b"
         ms["get_all_buffer_data"].return_value = "spectrums", "stats"
         ms["set_all_buffer_done"].return_value = False
-        assert interface.synchronized_poll_data() == (20, "spectrums", "stats")
+        assert interface.synchronized_poll_data(_acq_number) == (
+            20,
+            "spectrums",
+            "stats",
+        )
         ms["get_current_pixel"].assert_called_once_with()
         ms["any_buffer_overrun"].assert_called_once_with()
         assert ms["all_buffer_full"].call_args_list == [(("a",),), (("b",),)]
@@ -1142,21 +1149,31 @@ def test_get_handel_version(interface):
 
 
 def test_get_config_files(interface):
-    assert interface.get_config_files(".") == ["scripts/handel/mercury.ini"]
-    assert interface.get_config_files(b".") == [b"scripts/handel/mercury.ini"]
+    """ Test listing of config (.ini) files.
+    warning: Adding files in this directory can break the test.
+    """
+    assert interface.get_config_files("tests/mca/mercury_config_file") == [
+        "mercury.ini"
+    ]
+    assert interface.get_config_files(b"tests/mca/mercury_config_file") == [
+        b"mercury.ini"
+    ]
 
 
 def test_get_config(interface):
+    """ Test reading of configuration (.ini) files.
+    warning: Adding files in this directory can break the test.
+    """
     # Using string paths
-    filename = interface.get_config_files(".")[0]
-    conf = interface.get_config(".", filename)
+    filename = interface.get_config_files("tests/mca/mercury_config_file")[0]
+    conf = interface.get_config("tests/mca/mercury_config_file", filename)
     assert conf["detector definitions"][0]["alias"] == "detector1"
     with pytest.raises(IOError):
-        interface.get_config(".", "i_dont_exist.ini")
+        interface.get_config("tests/mca/mercury_config_file", "i_dont_exist.ini")
 
     # Using bytestring paths
-    filename = interface.get_config_files(b".")[0]
-    conf = interface.get_config(b".", filename)
+    filename = interface.get_config_files(b"tests/mca/mercury_config_file")[0]
+    conf = interface.get_config(b"tests/mca/mercury_config_file", filename)
     assert conf["detector definitions"][0]["alias"] == "detector1"
     with pytest.raises(IOError):
-        interface.get_config(b".", b"i_dont_exist.ini")
+        interface.get_config(b"tests/mca/mercury_config_file", b"i_dont_exist.ini")